feat: Mail Send, Mail Get 생성

+++ APIServer/Repository/IAuthRepository.cs
@@ -2,6 +2,6 @@ public interface IAuthRepository
 {
     // 취소 잘 처리하기
     Task<bool> CreateUserAsync(UserGameData data);
-    Task<bool> CheckUserAsync(int userId);
-    Task<UserGameData?> GetUserGameDataAsync(int userId);
+    Task<bool> CheckUserAsync(Int64 userId);
+    Task<UserGameData?> GetUserGameDataAsync(Int64 userId);
 }
