+++ HiveServer/Model/DTO/VerifyLogin.cs
@@ -1,6 +1,7 @@
+
 public class VerifyLoginReq
 {
-    public int UserId { get; set; }
+    public Int64 UserId { get; set; }
     public string Token { get; set; } = null!;
 }
