+++ HiveServer/Repository/IMemoryRepository.cs
@@ -1,6 +1,6 @@
 public interface IMemoryRepository
 {
-    public Task<string?> GetAccessToken(int user_id);
-    public Task<bool> SetAccessToken(int user_id, string token);
-    public Task<bool> DeleteAccessToken(int user_id);
+    public Task<string?> GetAccessToken(long user_id);
+    public Task<bool> SetAccessToken(long user_id, string token);
+    public Task<bool> DeleteAccessToken(long user_id);
 }
